Criado novo service para verificar situação do cliente

+++ ms-gym/src/main/java/br/janioofi/msgym/domain/dtos/PlanoDTO.java
@@ -11,5 +11,5 @@ public record PlanoDTO (
     String descricao,
     @JsonFormat(pattern = "dd/MM/yyyy") LocalDate vigencia,
     BigDecimal preco,
-    Integer quantidadeMeses
+    Long quantidadeMeses
 ){}
